Finally got the classes to work.
Biggest thing was that I was calling the class directly and trying to access the method from there. This can not be done as there is no instance of the class to begin with so I needed to create the buyUpgrade function to create an instance of the class first with the parameters already fed in so that there was an instance of the class that would look something like this:

Upgrades(5, clickPower, add, 1)

Without the instance it is basically calling nothing.